SVG_LogoGenerator/SVGRepresentation/deepsvg/requirements.txt
pytorch==1.7.1
torchvision==0.8.2
numpy==1.16.1
matplotlib
svgwrite
cairosvg
IPython
six>=1.12.0
tensorflow
tensorboardX
pillow
svglib
tqdm
jupyter
scikit-image
pandas
networkx
moviepy
numba
sklearn
umap-learn
umap-learn[plot]
shapely
kivy